build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@41762 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/testcase/tstLdrDisasmTest.cpp
@@ -88,7 +88,7 @@ static DECLCALLBACK(int) DisasmTest1ReadCode(PDISCPUSTATE pDis, uint8_t offInstr
     if (cb + pDis->uInstrAddr + offInstr > sizeof(g_ab32BitCode))
         cb = cbMinRead;
     memcpy(&pDis->abInstr[offInstr], &g_ab32BitCode[pDis->uInstrAddr + offInstr], cb);
-    pDis->cbCachedInstr = offInstr + cb;
+    pDis->cbCachedInstr = offInstr + (uint8_t)cb;
     return VINF_SUCCESS;
 }
